Extract log view
